v1.63 - Modbus: Expanded uptime to 32-bit
CHANGED:
- Uptime register expanded from 16-bit to 32-bit (prevents 18-hour rollover)
- Now supports ~136 years maximum uptime
- Uptime spans registers 2-3 (low word/high word)
- All subsequent holding registers shifted by one (now at addresses 4-12)
- Total holding registers increased from 12 to 13

IMPROVED:
- Display shows uptime intelligently: seconds (<1h), hours (1h-1d), days (1d+)
- Optimized display space while maintaining readability

UPDATED:
- Web interface displays 32-bit uptime value correctly
- README documentation with corrected register addresses
- Python examples updated to read 13 registers
- Register table and usage notes updated

+++ src/modbus_handler.h
@@ -11,7 +11,7 @@
 struct HoldingRegisters {
     uint16_t sequential_counter;
     uint16_t random_number;
-    uint16_t uptime_seconds;
+    uint32_t uptime_seconds;  // 32-bit for ~136 years (spans registers 2-3)
     uint16_t free_heap_kb_low;
     uint16_t free_heap_kb_high;
     uint16_t min_heap_kb;
